+++ src/FreshBooksClient.php
@@ -16,6 +16,7 @@ use Psr\Http\Message\RequestFactoryInterface;
 use Psr\Http\Message\StreamFactoryInterface;
 use amcintosh\FreshBooks\Model\Client;
 use amcintosh\FreshBooks\Model\ClientList;
+use amcintosh\FreshBooks\Model\Identity;
 use amcintosh\FreshBooks\Model\Invoice;
 use amcintosh\FreshBooks\Model\InvoiceList;
 use amcintosh\FreshBooks\Model\Payment;
@@ -23,6 +24,7 @@ use amcintosh\FreshBooks\Model\PaymentList;
 use amcintosh\FreshBooks\Model\Tax;
 use amcintosh\FreshBooks\Model\TaxList;
 use amcintosh\FreshBooks\Resource\AccountingResource;
+use amcintosh\FreshBooks\Resource\AuthResource;
 
 class FreshBooksClient
 {
@@ -69,6 +71,16 @@ class FreshBooksClient
         );
     }
 
+    /**
+     * The identity details of the currently authenticated user.
+     *
+     * @link https://www.freshbooks.com/api/me_endpoint
+     * @return Identity
+     */
+    public function currentUser(): Identity
+    {
+        return (new AuthResource($this->httpClient))->getMeEndpoint();
+    }
 
     /**
      * FreshBooks clients resource with calls to get, list, create, update, delete