+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add current user and identity model (identity, business, business membership)
+
 ## 0.1.0
 
 Basic Invoicing workflow functionality